requirements.txt
# Web
streamlit>=1.35.0

# Data
pandas>=2.2.0
numpy>=1.26.0

# Machine Learning
scikit-learn>=1.4.0
lightgbm>=4.3.0

# Deep Learning
torch>=2.3.0

# Database
mysql-connector-python>=8.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.22.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.4.0
